requirements.txt
ujson
googletrans==4.0.0rc1
pyTelegramBotAPI
emoji
requests
chardet==3.*
sqlalchemy
psycopg2
igdb-api-v4